+++ DMCompiler/DM/DMProc.cs
@@ -7,6 +7,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using DMCompiler.Bytecode;
 using OpenDreamShared.Compiler;
 
 namespace DMCompiler.DM {
