+++ DMCompiler/DM/Expressions/LValue.cs
@@ -1,7 +1,7 @@
 using System.Diagnostics.CodeAnalysis;
+using DMCompiler.Bytecode;
 using OpenDreamShared.Compiler;
 using OpenDreamShared.Dream;
-using OpenDreamShared.Dream.Procs;
 
 namespace DMCompiler.DM.Expressions {
     abstract class LValue : DMExpression {
