+++ DMCompiler/DM/Visitors/DMVisitorExpression.cs
@@ -3,7 +3,6 @@ using DMCompiler.DM.Expressions;
 using OpenDreamShared.Compiler;
 using DMCompiler.Compiler.DM;
 using OpenDreamShared.Dream;
-using OpenDreamShared.Dream.Procs;
 using Robust.Shared.Utility;
 
 namespace DMCompiler.DM.Visitors {
