+++ OpenDreamClient/Interface/Prompts/AlertWindow.cs
@@ -1,5 +1,5 @@
-﻿using OpenDreamShared.Dream.Procs;
-using JetBrains.Annotations;
+﻿using JetBrains.Annotations;
+using OpenDreamShared.Dream;
 using Robust.Shared.Console;
 
 namespace OpenDreamClient.Interface.Prompts;
