+++ OpenDreamClient/Interface/Prompts/TextPrompt.cs
@@ -1,4 +1,4 @@
-﻿using OpenDreamShared.Dream.Procs;
+﻿using OpenDreamShared.Dream;
 using Robust.Client.UserInterface.Controls;
 
 namespace OpenDreamClient.Interface.Prompts;
